requirements.txt
Flask~=2.1.2
WTForms~=3.0.1
Werkzeug~=2.1.2
numpy~=1.22.3
Pillow~=9.1.1
scikit-learn~=1.1.0